refactor(all): change filenames, add maps, and add errormsg for bigmap
