+++ maps/invalid/34_unexpected_char.cub
@@ -1,0 +1,13 @@
+42Tokyo
+
+NO ./textures/north.xpm
+SO ./textures/south.xpm
+WE ./textures/west.xpm
+EA ./textures/east.xpm
+
+F 220,100,0
+C 225,30,0
+
+111
+1E1
+111
