+++ src/calc_texture.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   tex2.c                                             :+:      :+:    :+:   */
+/*   calc_texture.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/07 14:43:04 by atashiro          #+#    #+#             */
-/*   Updated: 2026/02/07 14:48:42 by atashiro         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:12:50 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
