+++ src/calc_wall_height.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   calcurate_wall_height.c                            :+:      :+:    :+:   */
+/*   calc_wall_height.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/28 06:09:11 by atashiro          #+#    #+#             */
-/*   Updated: 2026/02/01 19:37:54 by atashiro         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:13:02 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
