+++ src/print_config_error.c
@@ -1,47 +1,41 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   error_map_print.c                                  :+:      :+:    :+:   */
+/*   print_config_error.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sohyamaz <sohyamaz@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/02/18 14:23:25 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/18 20:00:20 by sohyamaz         ###   ########.fr       */
+/*   Created: 2026/02/18 01:41:00 by sohyamaz          #+#    #+#             */
+/*   Updated: 2026/02/23 14:14:43 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3D.h"
 
-void	print_invalid_symbol(void)
+void	print_invalid_wall(void)
 {
 	print_prefix();
-	printf("There is at least one error symbol in your map\n");
+	printf("Wall setting is invalid\n");
 	return ;
 }
 
-void	print_player_error(int player)
+void	print_invalid_color(void)
 {
 	print_prefix();
-	if (player == 0)
-		printf("There is no player in your map\n");
-	else
-		printf("There are two or more player in your map\n");
+	printf("Color of floor/Ceiling is invalid\n");
 	return ;
 }
 
-void	print_torned_floor(void)
+void	print_unexpected_char(void)
 {
 	print_prefix();
-	printf("There is at least one hole on the floor in your map\n");
+	printf("Unexpected charactor detected\n");
 	return ;
 }
 
-void	print_wall_error(char c)
+void	print_dup_config_err(void)
 {
 	print_prefix();
-	if (is_user(c) == true)
-		printf("User will sporn at out of range\n");
-	else
-		printf("This map is not surrounded by wall\n");
+	printf("This map has duplicated congfig contents\n");
 	return ;
 }
