+++ src/print_error.c
@@ -1,34 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   error_config_print.c                               :+:      :+:    :+:   */
+/*   print_error.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sohyamaz <sohyamaz@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/02/18 01:41:00 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/18 03:38:02 by sohyamaz         ###   ########.fr       */
+/*   Created: 2026/02/17 23:52:50 by sohyamaz          #+#    #+#             */
+/*   Updated: 2026/02/23 14:14:49 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3D.h"
 
-void	print_invalid_wall(void)
+void	print_prefix(void)
 {
-	print_prefix();
-	printf("Wall setting is invalid\n");
+	printf("Error\n");
 	return ;
 }
 
-void	print_invalid_color(void)
+void	print_invalid_extention(void)
 {
 	print_prefix();
-	printf("Color of floor/Ceiling is invalid\n");
-	return ;
-}
-
-void	print_dup_config_err(void)
-{
-	print_prefix();
-	printf("This map has duplicated congfig contents\n");
+	printf("Extention is not valid\n");
+	printf("Give me a file with .cub extention\n");
 	return ;
 }
