+++ src/print_mapdata_error.c
@@ -1,49 +1,47 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   error_print.c                                      :+:      :+:    :+:   */
+/*   print_mapdata_error.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sohyamaz <sohyamaz@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/02/17 23:52:50 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/18 15:22:27 by sohyamaz         ###   ########.fr       */
+/*   Created: 2026/02/18 14:23:25 by sohyamaz          #+#    #+#             */
+/*   Updated: 2026/02/23 14:15:08 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3D.h"
 
-void	print_prefix(void)
+void	print_invalid_symbol(void)
 {
-	printf("Error\n");
+	print_prefix();
+	printf("There is at least one error symbol in your map\n");
 	return ;
 }
 
-void	print_arg_error(void)
+void	print_player_error(int player)
 {
 	print_prefix();
-	printf("Given args are not valid\n");
-	printf("Usage: ./game path_of_the_config_file.cub\n");
+	if (player == 0)
+		printf("There is no player in your map\n");
+	else
+		printf("There are two or more player in your map\n");
 	return ;
 }
 
-void	print_lack_of_config(void)
+void	print_torned_floor(void)
 {
 	print_prefix();
-	printf("There is not enough configure infomation\n");
+	printf("There is at least one hole on the floor in your map\n");
 	return ;
 }
 
-void	print_invalid_extention(void)
+void	print_wall_error(char c)
 {
 	print_prefix();
-	printf("Extention is not valid\n");
-	printf("Give me a file with .cub extention\n");
-	return ;
-}
-
-void	print_nomap_error(void)
-{
-	print_prefix();
-	printf("There is no map in .cub file\n");
+	if (is_user(c) == true)
+		printf("User will sporn at out of range\n");
+	else
+		printf("This map is not surrounded by wall\n");
 	return ;
 }
