--- src/raycasting_utils.c
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   raycasting_utils.c                                 :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/01/26 01:04:20 by atashiro          #+#    #+#             */
-/*   Updated: 2026/01/31 19:02:53 by atashiro         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../include/cub3D.h"
-
-float	distance(float x, float y)
-{
-	return (sqrt(x * x + y * y));
-}
