+++ src/texture.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   tex.c                                              :+:      :+:    :+:   */
+/*   texture.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/31 19:43:32 by atashiro          #+#    #+#             */
-/*   Updated: 2026/02/19 01:13:57 by sohyamaz         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:15:56 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
