+++ src/utils_for_minimap.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minimap_utils.c                                    :+:      :+:    :+:   */
+/*   utils_for_minimap.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/31 21:26:21 by atashiro          #+#    #+#             */
-/*   Updated: 2026/02/19 17:50:47 by atashiro         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:16:31 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
