+++ src/utils_for_parsing.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_util.c                                       :+:      :+:    :+:   */
+/*   utils_for_parsing.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sohyamaz <sohyamaz@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/09 05:40:29 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/13 02:27:07 by sohyamaz         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:16:36 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
