+++ src/utils_for_parsing_map.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_map_util.c                                   :+:      :+:    :+:   */
+/*   utils_for_parsing_map.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: atashiro <atashiro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/09 05:39:51 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/21 14:26:52 by atashiro         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:16:43 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
