+++ src/utils_for_read.c
@@ -1,43 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   init_util.c                                        :+:      :+:    :+:   */
+/*   utils_for_read.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sohyamaz <sohyamaz@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/01 19:41:27 by sohyamaz          #+#    #+#             */
-/*   Updated: 2026/02/17 04:35:39 by sohyamaz         ###   ########.fr       */
+/*   Updated: 2026/02/23 14:17:20 by sohyamaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3D.h"
-//
-//bool	get_map_size(t_map *map, char *fullpath)
-//{
-//	size_t	x;
-//	char	*line;
-//
-//	if (map == NULL || fullpath == NULL)
-//		return (false);
-//	map->cubfd = open(fullpath, O_RDONLY);
-//	if (map->cubfd < 0)
-//		return (perror("open"), false);
-//	x = 0;
-//	while (1)
-//	{
-//		line = get_next_line(map->cubfd);
-//		if (line == NULL)
-//			break ;
-//		x = ft_strlen(line) + 1;
-//		if (x > map->x)
-//			map->x = x;
-//		map->y++;
-//		free(line);
-//	}
-//	map->y++;
-//	close(map->cubfd);
-//	return (true);
-//}
 
 bool	is_config_line(char c)
 {
